requirements.txt
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==21.9b0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.6
click==8.0.3
decorator==5.1.0
distlib==0.3.3
Django==3.2.8
djangorestframework==3.12.4
Faker==9.2.0
filelock==3.3.0
identify==2.3.0
idna==3.2
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.28.0
isort==5.9.3
jedi==0.18.0
matplotlib-inline==0.1.3
mixer==7.1.2
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.15.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2021.10.8
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.8.1
wcwidth==0.2.5
